requirements.txt
telethon==1.37.0
python-telegram-bot==21.9
python-dotenv==1.1.0
aiohttp==3.11.12
pyyaml==6.0.2
